+++ .gitignore
@@ -23,3 +23,5 @@ assets
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+roadmap.md
